Fix camel-case pattern for matching schema field names.
Note: This is a better pattern, but it does not solve the fundamental ambiguity
between CamelCase and UPPER_CASE variations of the same name. 'Cvr' and 'CVR'
are still ambiguous.

It's possible to subclass the model for CVRs and have it know that fields
starting with "cvr" should be treated specially. But it seems cleaner to just
use the 'alias' keyword on 'Field'.